Remove exception if not root and port < 1024
This is unnecessary as an exception will be raised in just a few lines ('Could not use port {0}, {1}') if the user doesn't have access. This code prevents non-root users who DO have permissions to listen on low numbered ports from doing so. Such as users with an authbind permission. So I recommend removing it.